+++ include/swift/Runtime/ImageInspectionCommon.h
