+++ include/swift/Runtime/SymbolInfo.h
