+++ stdlib/public/runtime/ImageInspectionWasm.cpp
@@ -18,8 +18,8 @@
 
 #if defined(__wasm__)
 
+#include "swift/Runtime/ImageInspection.h"
 #include "swift/shims/MetadataSections.h"
-#include "ImageInspection.h"
 
 using namespace swift;
